bug 794261 - duplicates cron app moved into matviews
