+++ config/crontabber.ini
@@ -9,7 +9,7 @@
             socorro.cron.jobs.matviews.ADUCronApp|1d|02:00
             socorro.cron.jobs.matviews.HangReportCronApp|1d|02:00
             socorro.cron.jobs.matviews.NightlyBuildsCronApp|1d|02:00
-            socorro.cron.jobs.duplicates.DuplicatesCronApp|1h
+            socorro.cron.jobs.matviews.DuplicatesCronApp|1h
             socorro.cron.jobs.matviews.ReportsCleanCronApp|1h
             socorro.cron.jobs.bugzilla.BugzillaCronApp|1h
             socorro.cron.jobs.matviews.BuildADUCronApp|1d|02:00
