+++ socorro/cron/jobs/matviews.py
@@ -156,3 +156,19 @@ class ReportsCleanCronApp(PostgresBackfillCronApp, _Base):
     def run(self, connection, date):
         date -= datetime.timedelta(hours=2)
         self.run_proc(connection, [date])
+
+
+class DuplicatesCronApp(PostgresBackfillCronApp, _Base):
+    proc_name = 'update_reports_duplicates'
+    app_name = 'duplicates'
+    app_version = '1.0'
+    app_description = ""
+
+    def run(self, connection, date):
+        start_time = date - datetime.timedelta(hours=3)
+        end_time = start_time + datetime.timedelta(hours=1)
+        self.run_proc(connection, [start_time, end_time])
+
+        start_time += datetime.timedelta(minutes=30)
+        end_time = start_time + datetime.timedelta(hours=1)
+        self.run_proc(connection, [start_time, end_time])
